requirements.txt
pyyaml
flask
flask-restful
colorama
pytest
pandas
bs4